admin/requirements.txt
Django
djangorestframework
mysqlclient 
django-mysql
django-cors-headers
kafka-python